requirements.txt
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=0.19.0
google-genai>=0.3.0
pandas>=2.0.0
matplotlib>=3.7.0
notebook>=7.0.0
